userDetail 영상 업데이트, deleteVideo 보안

+++ controller/videoController.js
@@ -52,7 +52,7 @@ export const videoDetail = async (req, res) => {
 
   try {
     const video = await Video.findById(id).populate("creator"); //populate 객채를 불러오는 함수 오직 objectID 타입에만 사용 가능
-    console.log(video);
+    // console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch(error){
     console.log(error);
@@ -60,13 +60,17 @@ export const videoDetail = async (req, res) => {
   }
 }
 
-export const getEditVideo = async (req, res) => {
+export const getEditVideo = async (req, res) => { //크리에이터 id, 로그인 id가 다르면 edit 할 수 없어야 함
   const {
     params: {id} //파라미터
   } = req;
   try{
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video});
+    if(video.creator !== req.user.id){ //creator도 id
+      throw Error(); //try 내에서 에러 발생 시 자동으로 catch로 이동
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video});
+    }
   } catch(error){
     res.redirect(routes.home);
   }
@@ -91,9 +95,14 @@ export const deleteVideo = async (req, res) =>{
     params: {id}
   } = req;
   try{
-    await Video.findOneAndRemove({_id: id});
+    const video = await Video.findById(id);
+    if(video.creator !== req.user.id){
+      throw Error();
+    } else{
+      await Video.findOneAndRemove({_id: id});
+    }
   } catch(error){
-    
+    console.log(error);
   }
   res.redirect(routes.home)
 }
